+++ src/lib/pages/public-project/index.tsx
@@ -4,19 +4,15 @@ import PageContainer from "lib/components/PageContainer";
 
 import { AllProject } from "./components/AllProject";
 
-const AllPublicProjectsPage = () => {
-  return (
-    <PageContainer>
-      <Flex direction="column" alignItems="center" gap={8}>
-        <Flex justifyContent="space-between" w="full" alignItems="center">
-          <Heading as="h5" variant="h5" color="primary.400">
-            Public Projects
-          </Heading>
-        </Flex>
-        <AllProject />
+export const AllPublicProjectsPage = () => (
+  <PageContainer>
+    <Flex direction="column" alignItems="center" gap={8}>
+      <Flex justifyContent="space-between" w="full" alignItems="center">
+        <Heading as="h5" variant="h5" color="primary.400">
+          Public Projects
+        </Heading>
       </Flex>
-    </PageContainer>
-  );
-};
-
-export default AllPublicProjectsPage;
+      <AllProject />
+    </Flex>
+  </PageContainer>
+);